separate create_image_lists into utils
